Create 'sched' module for scheduling tasks to run after a timeout
There will be a queue of tasks each specifying a timeout and a callback
to run after the timeout has occured. A task can be added to the queue
with 'sched_add' and removed with 'sched_rem'.

Tasks are automatically removed from the queue once they have executed
unless the callback returns true at which point the task is re-scheduled
with the same timeout.