Add code for adding, removing and processing tasks in the scheduler
When a task is added the timeout is added onto the current time to
calculate the time at which the callback should be called. The ISR
checks the current time against this calculated future time and runs
the callback if it's due.

There should be no problems with overflow of the global time as both
the code and time will overflow in the same way when the timeout is
added onto the current time and when the timer reaches the maximum count.