Change piezo module to maintain the note buffer internally
Rather than using the callback to generate the next note for playback it
acts as a notification that the internal buffer is running low on notes.

To play a tune an array of notes is passed to the piezo_play function
which manages filling the buffer and ensuring that it doesn't overflow.
The piezo_play function allows for simple tunes to be played by passing
an array and a length and more complex ones to be played, i.e. from the
beagleboard, by passing some note data, a length and true for 'notif'.

When more data is required the buf_low callback is executed which allows
what ever mechanism is being used to get note data to load more data into
the buffer through the use of the piezo_play function.