Restructure the charger detection, don't rely solely on the pluging action
The charger can be plugged in but not turned on so the interrupt generated
when it's plugged in isn't much use in itself. There's now a function which
checks to see if the charger is present and if it is actually charging the
battery. This function is called every 5 seconds to account for the charger
being turned off but left plugged in. It is also called when the charger is
plugged in or removed to ensure a nice and snappy response with the sound fx.