Don't think the battery's charging when the charger isn't plugged in
Oops, forgot to take into account the state of the charger plug switch.
If the charger isn't plugged in then there's no way it can be charging.
All of the other logic still holds for when it is plugged in though as
it could be connected but switched off at the wall.