Add a function to check if the BeagleBoard is powered on
The 'REGEN' line is from the PMIC on the BeagleBoard. It is actually meant to
be for controlling regulators and things on an expansion board, ensuring that
they're only turned on after the BB has turned on. It works well as an
indication of the BB being turned on and running.