Check the current state of the BeagleBoard power during initialisation
If the BealgeBoard is already powered when the MSP boots then we want to leave
it on as the boot was caused either by a FET or a firmware upgrade.
If the BeagleBoard isn't already powered then the power board must have just
been turned on so we want to hold off on powering the BeagleBoard until some
POST stuff has been completed.