Remove erroneous assumption that piezo_t is the same size as the data it holds
sric_play_piezo got the input data size comparison wrong, because the
size of the struct is not 5, it's at least 6 due to padding.

This patch also removes assumptions about the endianness of the host
and msp.  msp430-libc doesn't contain htons or anything like that, so
we have to code it manually.  This way, there is absolutely no
ambiguity about byte ordering.