Setup and use the Supply Voltage Supervisor
Reset the MSP if the supply voltage drops below 2.5V

When testing the voltage rail monitoring code I experienced random
failure where the board would get stuck in a state where the motor
rail was on even though the supply voltage was below 7V. This problem
occured more and more as the battery got flatter. After a few hours of
debugging through LEDs and then monitoring the 3.3V rail I have come
to the conclusion that registers are being muddled as the supply to the
MSP was dipping to about 1V. This now forces the MSP to reset under
these extreme circumstances.