Ensure that the charger detect interrupt edge select is set properly
When setting the edge sensitivity for the charger detect pin interrupt
the flag was being cleared and then set. This causes an erroneous
interrupt leading to an endless loop of the board thinking a charger
was being plugged in. Once one had been plugged in for the first time it
went away.