Re-enable the motor rail once the battery voltage reaches 12V
If the battery is flat and the motor rail has been disabled then
the only way of re-enabling it, other than running some new code,
is to charge the battery up. This checks the voltage on the battery
when the charger isn't present to see if it's charged enough yet.