Add a gating mechanism to the motor rail enable function
The motor rail will only be enabled when both the charger detection code
and pyenv say it can be enabled. If either thing disables the motor rail
then it will be disabled straight away.

The charger detection code will disable the motor rail as soon as a charger
is plugged in irrespective of the charger being switched on.