Alter the battery cut-off threshold
The previous threshold was too low and almost completely drained the LiPo
batteries. As the power board draws a very small current when it's
switched off some batteries were getting over-discharged.

This increases the threshold to 3.4V per cell, which is at the end of the
linear section of the discharge curve. The battery should last for
~30 days if left connected to a switched-off power board after the power
board has indicated the battery is flat.

This also adds a second beep on boot so that we can confirm the prescence
of the new firmware.