Remove the motor rail charge gating code
A side effect of removing the charger handling code was that the motor
rail was never being enabled. The gating code ensured that the motor
rail was only ever turned on if both the charger was unplugged and the
BeagleBoard had commanded it to do so. As we no longer use the charger
this code is now useless.

Saves 2 bytes in .data